Extract MainAppBar
