+++ lib/di.config.dart
@@ -1,5 +1,5 @@
-// dart format width=80
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// dart format width=80
 
 // **************************************************************************
 // InjectableConfigGenerator
@@ -43,7 +43,7 @@ extension GetItInjectableX on _i174.GetIt {
   }) {
     final gh = _i526.GetItHelper(this, environment, environmentFilter);
     final extrernalModule = _$ExtrernalModule();
-    gh.singleton<_i403.AppService>(() => _i403.AppService.new());
+    gh.singleton<_i403.AppService>(() => _i403.AppService());
     gh.singleton<_i429.EncryptionKeyRegistry>(
       () => _i429.EncryptionKeyRegistry(),
     );
